+++ src/test/java/org/cicirello/ds/BinaryHeapTests.java
@@ -797,127 +797,6 @@ public class BinaryHeapTests {
 	// MAX HEAP TESTS
 	
 	@Test
-	public void testDefaultMaxHeap() {
-		int n = 31;
-		String[] elements = createStringsMaxCase(n);
-		int[] priorities = createPriorities(elements);
-		PriorityQueueNode.Integer<String>[] pairs = createPairs(elements, priorities);
-		BinaryHeap<String> pq = BinaryHeap.createMaxHeap();
-		assertEquals(0, pq.size());
-		assertTrue(pq.isEmpty());
-		assertNull(pq.peekElement());
-		assertNull(pq.peek());
-		assertEquals(Integer.MIN_VALUE, pq.peekPriority());
-		for (int i = 0; i < n; i++) {
-			assertTrue(pq.offer(pairs[i]));
-			assertEquals(i+1, pq.size());
-			assertFalse(pq.isEmpty());
-			assertEquals("A", pq.peekElement());
-			assertEquals(pairs[0], pq.peek());
-			assertEquals((int)'A', pq.peekPriority());
-		}
-		for (int i = 0; i < n; i++) {
-			assertTrue(pq.contains(elements[i]));
-			assertTrue(pq.contains(pairs[i]));
-			assertFalse(pq.offer(pairs[i]));
-			assertEquals(n, pq.size());
-			assertEquals("A", pq.peekElement());
-			assertEquals(pairs[0], pq.peek());
-			assertEquals((int)'A', pq.peekPriority());
-		}
-		for (int i = 0; i < n; i++) {
-			assertEquals(priorities[i], pq.peekPriority(elements[i]));
-		}
-		assertEquals(Integer.MIN_VALUE, pq.peekPriority("hello"));
-		for (int i = 0; i < n; i++) {
-			assertEquals(pairs[i], pq.poll());
-			assertFalse(pq.contains(pairs[i].element));
-			assertEquals(n-1-i, pq.size());
-		}
-		assertNull(pq.poll());
-	}
-	
-	@Test
-	public void testDefaultMaxHeapReverse() {
-		int n = 31;
-		String[] elements = createStringsRevMaxCase(n);
-		int[] priorities = createPriorities(elements);
-		PriorityQueueNode.Integer<String>[] pairs = createPairs(elements, priorities);
-		BinaryHeap<String> pq = BinaryHeap.createMaxHeap();
-		assertEquals(0, pq.size());
-		assertTrue(pq.isEmpty());
-		assertNull(pq.peekElement());
-		assertNull(pq.peek());
-		assertEquals(Integer.MIN_VALUE, pq.peekPriority());
-		for (int i = 0; i < n; i++) {
-			assertTrue(pq.offer(pairs[i]));
-			assertEquals(i+1, pq.size());
-			assertFalse(pq.isEmpty());
-			assertEquals(elements[i], pq.peekElement());
-			assertEquals(pairs[i], pq.peek());
-			assertEquals((int)elements[i].charAt(0), pq.peekPriority());
-		}
-		for (int i = 0; i < n; i++) {
-			assertTrue(pq.contains(elements[i]));
-			assertTrue(pq.contains(pairs[i]));
-			assertFalse(pq.offer(pairs[i]));
-			assertEquals(n, pq.size());
-			assertEquals("A", pq.peekElement());
-			assertEquals(pairs[n-1], pq.peek());
-			assertEquals((int)'A', pq.peekPriority());
-		}
-		for (int i = 0; i < n; i++) {
-			assertEquals(priorities[i], pq.peekPriority(elements[i]));
-		}
-		assertEquals(Integer.MIN_VALUE, pq.peekPriority("hello"));
-		for (int i = 0; i < n; i++) {
-			assertEquals(pairs[n-1-i], pq.poll());
-			assertFalse(pq.contains(elements[n-1-i]));
-			assertEquals(n-1-i, pq.size());
-		}
-		assertNull(pq.poll());
-	}
-	
-	@Test
-	public void testDefaultMaxHeapArbitrary() {
-		int n = 31;
-		String[] elements = createStringsArbitraryMaxCase(n);
-		int[] priorities = createPriorities(elements);
-		PriorityQueueNode.Integer<String>[] pairs = createPairs(elements, priorities);
-		BinaryHeap<String> pq = BinaryHeap.createMaxHeap();
-		assertEquals(0, pq.size());
-		assertTrue(pq.isEmpty());
-		assertNull(pq.peekElement());
-		assertNull(pq.peek());
-		assertEquals(Integer.MIN_VALUE, pq.peekPriority());
-		for (int i = 0; i < n; i++) {
-			assertTrue(pq.offer(pairs[i]));
-			assertEquals(i+1, pq.size());
-			assertFalse(pq.isEmpty());
-		}
-		for (int i = 0; i < n; i++) {
-			assertTrue(pq.contains(elements[i]));
-			assertTrue(pq.contains(pairs[i]));
-			assertFalse(pq.offer(pairs[i]));
-			assertEquals(n, pq.size());
-			assertEquals("A", pq.peekElement());
-			assertEquals(new PriorityQueueNode.Integer<String>("A",(int)'A'), pq.peek());
-			assertEquals((int)'A', pq.peekPriority());
-		}
-		for (int i = 0; i < n; i++) {
-			assertEquals(priorities[i], pq.peekPriority(elements[i]));
-		}
-		assertEquals(Integer.MIN_VALUE, pq.peekPriority("hello"));
-		for (int i = 0; i < n; i++) {
-			String expected = ""+((char)('A'-i));
-			assertEquals(new PriorityQueueNode.Integer<String>(expected, (int)('A'-i)), pq.poll());
-			assertFalse(pq.contains(expected));
-			assertEquals(n-1-i, pq.size());
-		}
-		assertNull(pq.poll());
-	}
-	
-	@Test
 	public void testListMaxHeap() {
 		int n = 31;
 		String[] elements = createStringsMaxCase(n);
